chore: Reusable app window components
- Removes deprecated dialog implementation
